상품 상세 페이지 sidebar 및 navbar 이동 구현
