+++ package.json
@@ -22,6 +22,7 @@
     "react-redux": "^7.2.6",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
+    "react-scroll": "1.8.4",
     "redux": "^4.1.2",
     "styled-components": "^5.3.3",
     "web-vitals": "^2.1.2"
