+++ src/pages/Product/components/Ask.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import styled from 'styled-components'
+
+const Ask = (props) => {
+  return <Container id={props.id}>ddd</Container>
+}
+
+const Container = styled.div`
+  display: flex;
+`
+
+export default Ask
