+++ src/pages/Product/components/Delivery.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import styled from 'styled-components'
+
+const Delivery = (props) => {
+  return <Container id={props.id}>ddd</Container>
+}
+
+const Container = styled.div`
+  display: flex;
+`
+
+export default Delivery
