+++ src/pages/Product/components/ProductNavigation.js
@@ -1,9 +1,10 @@
 import React, { useState } from 'react'
 import styled from 'styled-components'
 import { NavbarElement } from '../elements'
+import { Link } from 'react-scroll'
 
 const element = ['상품정보', '리뷰', '문의', '배송/환불', '추천']
-
+const linkTo = ['productInfo', 'review', 'ask', 'delivery', 'recommend']
 const ProductNavigation = () => {
   const [selectedElem, setSelectedElem] = useState(null) //선택된 navbar element의 id
 
@@ -16,14 +17,17 @@ const ProductNavigation = () => {
     <Container>
       <Nav>
         {element.map((data, idx) => (
-          <NavbarElement
-            id={idx}
-            text={data}
-            onClick={() => {
-              onNavbarSelect(idx)
-            }}
-            selectedElem={selectedElem}
-          />
+          //해당 id를 가진 컴포넌트로 이동 후 -70만큼 padding(navbar에 가려짐)
+          <Link style={{ width: '20%' }} to={linkTo[idx]} offset={-70}>
+            <NavbarElement
+              id={idx}
+              text={data}
+              onClick={() => {
+                onNavbarSelect(idx)
+              }}
+              selectedElem={selectedElem}
+            />
+          </Link>
         ))}
       </Nav>
     </Container>
@@ -37,12 +41,15 @@ const Container = styled.nav`
   background-color: #fafafa;
   border: 1px solid #dddddd;
   width: 100%;
+  position: sticky;
+  top: 0px;
+  z-index: 1;
 `
 const Nav = styled.ol`
   display: flex;
   align-items: center;
   justify-content: space-between;
-  width: 80%;
+  width: 767px;
   font-size: 16px;
   font-weight: bold;
   margin: 0px;
