+++ src/pages/Product/components/Review.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import styled from 'styled-components'
+
+const Review = (props) => {
+  return <Container id={props.id}>ddd</Container>
+}
+
+const Container = styled.div`
+  display: flex;
+`
+
+export default Review
