+++ src/pages/Product/components/index.js
@@ -1,14 +1,28 @@
 import ProductSimpleView from './ProductSimpleView'
 import Option from './Option'
 import ProductNavigation from './ProductNavigation'
-import ProductInfo from './ProductInfo'
+import UserStylingView from './UserStylingView'
 import UserStylingPhoto from './UserStylingPhoto'
 import SelectedOption from './SelectedOption'
+import ProductInfo from './ProductInfo'
+import ProductSidebar from './ProductSidebar'
+import OptionSelectView from './OptionSelectView'
+import Review from './Review'
+import Ask from './Ask'
+import Delivery from './Delivery'
+import Recommend from './Recommend'
 export {
   ProductSimpleView,
   Option,
   ProductNavigation,
-  ProductInfo,
+  UserStylingView,
   UserStylingPhoto,
   SelectedOption,
+  ProductInfo,
+  ProductSidebar,
+  OptionSelectView,
+  Review,
+  Ask,
+  Delivery,
+  Recommend,
 }
